+++ Firmware/include/globals.h
@@ -46,6 +46,10 @@ extern int level_trim_portrait_pos_deci_deg;
 extern int level_trim_portrait_neg_deci_deg;
 extern int reticle_offset_x;
 extern int reticle_offset_y;
+extern bool brightness_auto;
+extern int brightness_manual_pct;
+extern int brightness_auto_top_pct;
+extern bool show_horizon_line;
 
 // Filter algorithm
 extern int samples[SMOOTHING_WINDOW_SIZE];